requirements.txt
PyQt5
fpdf
pillow
BeautifulSoup4
httpx[http2]